:tada: :sparkles: :truck: Moved Project
Moved the Firebase Cloud Functions code from the stegappasaurus project (which now called
stegappasaurus-app), to here. The old project is here https://gitlab.com/stegappasaurus/stegappasaurus-app
(where the code used to be). Both projects exist in the stegappasaurus group on GitLab.

The reason for the move was the code is fairly independent, the React Native app just uses as a web API.
They had two different tsconfigs and packagae.json for example. The GitLab CI was a bit messy as well,
needing to install. Far cleaner to have the FAAS in a separate module.
some deps ad dev deps in the main package.json.